Add ability to stream and record video to client.
The client can now get video data from the camera, store it as a raw video file, and then convert that raw data to a MP4 file once recording is finished. Additionally the client attempts to stream the video data to twitch using the livestream functionality, but is unsucessful since RTMP livestreaming does not work at the moment.